requirements.txt
beautifulsoup4==4.8.1
bootstrap4==0.1.0
certifi==2019.9.11
chardet==3.0.4
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.5
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.0
